Applying a patch helpfully provided by James Synge.  From his e-mail:
1) If the user fails to set the "first change" when creating a new
job, the plugin will attempt to report all changes since the creation
of the depot as applying to the current build.  When your depot has
~300,000 changes, this is a very expensive operation.  The attached
patch does not rely on the user to provide that value (if not
supplied, it defaults to zero, and that is assumed to mean "don't
report the diffs").

2) If the user sets the depot path to "//..." (thinking it means
everything selected by the p4 workspace), as a co-worker did twice
this week, then it the plugin executes "p4 changes -m 25 //...@N" for
values of N, slowly decending from the latest change (from 300,000 in
our case).  This operation is even worse than the previous one,
causing the p4 server to appear to lockup.  The patch doesn't rely on
the user to provide that path in order to get this right.

3) The code to determine changes and sync the workspace treated each
mapping in the viewspec as a separate item to be operated on.  This is
very expensive when the mapping ends with /* instead of /... (due to
some internal perforce index being optimized for the latter, but
definitely not the former).  The patch operates on the entire
workspace when doing these operations.



git-svn-id: https://hudson.dev.java.net/svn/hudson/trunk/hudson/plugins@19140 71c3de6d-444a-0410-be80-ed276b4c234a