[artf54096]
with collabnet hudson plugin installed, svn-tag was no longer identifying a subversion repository as "SubversionSCM".  This was due to the fact that collabnet plugin was including subversion.jar & promoted-build.jar, so loading our own SubversionSCM caused the "instanceof" check in svn-tag to fail.

so removed the "copy" commands from the pom.xml.

git-svn-id: https://hudson.dev.java.net/svn/hudson/trunk/hudson/plugins@29269 71c3de6d-444a-0410-be80-ed276b4c234a